feature: metrics performance improvements
